% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joya_no_kane.R
\name{synthesize_bell}
\alias{synthesize_bell}
\title{鐘の音を合成する}
\usage{
synthesize_bell(freq = 150, duration = 4, sample_rate = 44100)
}
\arguments{
\item{freq}{基本周波数}

\item{duration}{持続時間}

\item{sample_rate}{サンプリングレート}
}
\value{
audioSampleオブジェクト
}
\description{
鐘の音を合成する
}
\keyword{internal}
